Pulled out scss for a bit, implemented media and constants
_Why_?

we are standardizing around using postcss and css modules and using the
sass syntax is fairly confusing as we get setup. we can bring it back
later using precss if we really need some of the functionality

_How_?

imported css module values and created colors and breakpoint definitions
in the styles directory.

examples of how to use these are in HeroHeader/styles.css

_Side Effects_?

pulled out scss processor, so we can't create any more scss files. if we
need scss syntax, we can bring back precss.