Added success message to email capture forms
_Why_?

so that users could get some visual feedback of the data having been
collected

_How_?

created a new redux module that is nothing more than a simple k/v store
that we can namespace messages to.

added createMessage and a clearMessages functions for managing this
state.

connected both email forms to this module and set a timeout to clear
that message after 5 seconds.

_Side Effects_?

naw.