Simplify Creation of Accenture Header
_Why_?

We didn't actually need to use inheritance, just pass in a new logo to
the header when we need!

Also, we wired redux to handle this state change rather than a manual
check of the subdomain every time, this way when we are getting this
data from the api, we can attach it to the specific user and they will
always see the version of thrive that we want!

_How_?

* Update header to accept children and replace logo with the child
* update accenture header to use this new header
* update our selector to read from redux
* update appcontainer to register subdomain on mount

_Side Effects_?

nope