Refactor Header Paths and WhiteLabeling logic
_Why_?

So that we just import Header and don't think about it. We just grab a
header and move on with our lives

_How_?

put accenture header into Header directory
update index to handle all connections and exports
remove any cruft that these changes rendered unecessary

_Side Effects_?

prettier code. (that is highly subjective)